requirements.txt
beautifulsoup4==4.13.5
black==25.1.0
bs4==0.0.2
cachetools==5.5.2
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==3.0.1
click==8.2.1
colorzero==2.0
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
distro==1.8.0
font-fredoka-one==0.0.4
font-hanken-grotesk==0.0.2
font-intuitive==0.0.4
font-source-sans-pro==0.0.1
font-source-serif-pro==0.0.1
fonts==0.0.3
fonttools==4.59.1
future==1.0.0
geocoder==1.38.1
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gpiod==2.3.0
gpiodevice==0.0.5
gpiozero==2.0.1
httplib2==0.22.0
idna==3.3
inky==2.2.1
kiwisolver==1.4.9
lgpio==0.2.2.0
lxml==6.0.1
matplotlib==3.10.5
mypy_extensions==1.1.0
numpy==2.2.1
oauthlib==3.3.1
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pigpio==1.78
pillow==11.0.0
platformdirs==4.4.0
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodomex==3.11.0
pyparsing==3.2.3
python-apt==2.6.0
python-dateutil==2.9.0.post0
pytz==2025.2
ratelim==0.1.6
requests==2.28.1
requests-oauthlib==2.0.0
rpi-lgpio==0.6
rsa==4.9.1
seaborn==0.13.2
six==1.16.0
slacker==0.14.0
smbus2==0.4.2
soupsieve==2.7
spidev==3.5
ssh-import-id==5.10
toml==0.10.2
typing_extensions==4.15.0
tzdata==2025.2
Unidecode==1.4.0
uritemplate==4.2.0
urllib3==1.26.12
wikiquotes==1.4
